requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
openai
tiktoken
faiss-cpu

